pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=77.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "annolid"
version = "1.4.0"
authors = [
    { name = "Chen Yang", email = "healthonrails@gmail.com" },
]
description = "An annotation and instance segmentation-based multiple animal tracking and behavior analysis package."
readme = "README.md"
requires-python = ">=3.10"
license = "CC-BY-NC-4.0"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

dependencies = [
    # Avoid NumPy 2 ABI breakage with compiled deps.
    'numpy>=1.24',
    'numba>=0.60.0',
    'scipy>=1.5.2',
    'easydict>=1.9',
    'opencv-contrib-python>=4.1.2.30',
    'PyYAML>=5.3',
    'sounddevice>=0.4.6',
    'librosa>=0.10.1',
    'torch>=2.5.0',
    'torchvision>=0.20.0',
    'pandas>=1.1.3',
    'pycocotools>=2.0.2',
    'tensorboard>=2.3.0',
    'imgviz>=1.7.5',
    'imageio>=2.8.0',
    'tqdm>=4.64.0',
    'shapely>=2.0.2',
    'imageio-ffmpeg>=0.4.2',
    'qimage2ndarray>=1.8',
    'qtpy>=2.4.0',
    'tokenizers',
    'einops>=0.6',
    'omegaconf>=2.3.0',
    'hydra-core>=1.3.2',
    'timm>=0.9.7',
    'h5py>=3.10.0',
    'psutil>=5.9.8',
    'pyserial>=3.5',
    'pyzmq>=25.1.2',
    'ffpyplayer>=4.5.0',
    'tree-config>=0.6.0',
    # 'PyQt5==5.15.7',
    'Pillow>=9.3.0,<12.0',
    'chardet>=5.2.0',
    'transformers>=4.45.1',
    'ultralytics>=8.4.0',
    'lap',
    'huggingface-hub>=0.22.0',
    'tensorboard',
    'jsonschema>=4.20.0',
    # Optional dependency: install SAM / SAM-HQ separately (see README.md).
]

[project.optional-dependencies]
gui = [
    # Qt binding for the GUI (Annolid ships its own LabelMe-compatible UI bits).
    "PyQt5>=5.15.7",
]

sam3 = [
    "iopath>=0.1.9",
    "ftfy>=6.1.1",
]

image_editing = [
    "diffusers>=0.32.0",
    "accelerate>=0.27.0",
    "safetensors>=0.4.0",
]

text_to_speech = [
    "pocket-tts>=1.0.0",
    "pyqtwebengine",
    "pymupdf",
]

qwen3_embedding = [
    "transformers>=4.57.0",
    "qwen-vl-utils>=0.0.14",
]

[project.urls]
Homepage = "https://github.com/healthonrails/annolid"
Repository = "https://github.com/healthonrails/annolid"

[project.scripts]
annolid = "annolid.gui.launcher:main"
annolid-run = "annolid.engine.cli:main"

[tool.setuptools.packages.find]
# This automatically finds your packages, same as setuptools.find_packages()
# By default it looks in the root directory. If you use a `src` layout,
# you would change this to: where = ["src"]
where = ["."]
include = ["annolid*"]
exclude = ["build*", "dist*", "docs*", "demo*", "tests*"]

[tool.setuptools.package-data]
annolid = [
    "**/*.js",
    "**/*.png",
    "**/*.jpg",
    "**/*.jpeg",
    "**/*.gif",
    "**/*.svg",
    "**/*.ico",
    "**/*.icns",
    "**/*.qrc",
    "core/output/schema/*.json",
]

[tool.ruff]
# Keep linting focused on source code; tutorial notebooks are authored for
# interactive execution and intentionally violate several static rules.
extend-exclude = [
    "book/tutorials/*.ipynb",
    "docs/tutorials/*.ipynb",
    # Vendored external projects keep upstream style/structure.
    "annolid/detector/countgd/**",
    "annolid/segmentation/cutie_vos/**",
    "annolid/segmentation/SAM/**",
    "annolid/depth/video_depth_anything/**",
]
